Set diffserv to CS6 for babel packets
Setting the IPv6 dscp field to CS6 allows for babel packets to
'jump' a bufferbloated queue and get to a destination sooner.

With the pfifo_fast qdisc (the default for ethernet), this
'does the right thing'.

With mac80211 wireless, this moves babel packets into the VO
queue, from which it is hopefully extracted and then multicast.