Remove "parasitic" mode.
This is the functionality described in Appendix C of RFC 6126.  Its
main purpose is to avoid keeping a full source table, which makes it
possible to implement a subset of Babel in just a few hundred lines of
code.  However, the code for maintaining the source table is already
there, and a parasitic implementation can be simulated using filtering
-- so it makes little sense to keep the functionality.