Fix interoperability with non-ss routers.
If we change the buffered destination prefix or router ID
in a ss-update TLV, then next update TLVs will depends on
it.  This cause some interoperability problems in some
cases, and can cause wrong messages in others.

The current solution should not be definitive: we should
send all non-source-specific update in group, but the
announce-with-default-source-prefix option introduce
non-source-specific updates.