Makefile: add support for whole unit compilation with full optimization
Babeld is a small program as things go. Compiling it as one unit
gives the compiler more chances to optimize.

babeld-whole compiled with -Os is ~2k smaller than babeld compiled
as separate object files.

With -O3 (the default for this option) whole unit compilation is
about 4k smaller. As for measuring a performance difference,
between -O3 and -Os, benchmarks await.

This commit also adds support for generating assembly language
output to make it easier to find further optimization opportunities.