Inline last_interface() into add_interface().
The check for `!ifp` in `last_interface()` is redundant with the check
`interface == NULL` in `add_interface()`. After eliminating it, it
becomes trivial to inline by hand the function `last_interface` which
is only used at this place. After inlining it, well, we can simply
replace the whole block with the standard way of manipulating
linked-lists in C.