Remove src_plen assignments which break v4mapped encoding
These assignments were introduced in f8bce04, probably
to ensure that src_plen doesn't contain any leftovers from
earlier revisions when source specific wasn't done via sub-TLVs.

This however breaks v4mapped encoding as a prefix length of 0
for IPv4 routes is represented as src_plen = 96. This offset is
already added outside of the affected functions and therefore
gets overwritten by this assignment.

Both functions contain an additional check if the message contains
ipv4 addresses, but as this check is only executed if the message has
a source-specific sub-TLV, the offset isn't added for non source-specific
routes.

Because the offset is already added outside of the subtlv parsing
functions, this additional check is also removed, as it would otherwise
break IPv4 source specifics by adding the IPv4 offset two times.

Signed-off-by: Fabian Bläse <fabian@blaese.de>