Fix interfaces leaking their if_conf.
To trigger this bug, launch babeld with a single interface and a dummy
configuration, and open a local socket.

    sudo valgrind --leak-check=full --show-leak-kinds=all -- \
        ./babeld -d2 -G$PWD/babeld.local -C 'interface eth0 unicast true' eth0

then, in the local socket, flush the interface, then kill babeld.

    5 bytes in 1 blocks are indirectly lost in loss record 2 of 6
       at 0x483A77F: malloc (vg_replace_malloc.c:307)
       by 0x4924B4E: strdup (in /usr/lib/libc-2.32.so)
       by 0x11C4AC: getword (configuration.c:111)
       by 0x11D686: getstring (configuration.c:129)
       by 0x11B6C6: parse_ifconf (configuration.c:682)
       by 0x11B6C6: parse_config_line (configuration.c:1057)
       by 0x11BE86: parse_config_from_string (configuration.c:1195)
       by 0x10A8B2: main (babeld.c:300)

    61 (56 direct, 5 indirect) bytes in 1 blocks are definitely lost in loss record 3 of 6
       at 0x483CB65: calloc (vg_replace_malloc.c:760)
       by 0x11B627: parse_ifconf (configuration.c:674)
       by 0x11B627: parse_config_line (configuration.c:1057)
       by 0x11BE86: parse_config_from_string (configuration.c:1195)
       by 0x10A8B2: main (babeld.c:300)

What happens is that the ifconf is added to the interface_confs list
in add_ifconf(), but in finalise_config(), the list is emptied and
interfaces are created with their if_conf. However, when flushing the
interface, it's the list that is searched, and the ifconf member of
the interface structure is not directely freed.

In case that there's another mechanism which I'm not aware of that
would allow an ifconf to be present in the list and in the interface
struct, we still search the list. If the ifconf wasn't found, we free
it. I think there's no way for interfaces to share an ifconf. We never
pass the default ifconf to the function when flushing an interface, so
we should be safe from use-after-free.