local: make local_kind function accessible
The kind-definitions are contained in the header file:
  LOCAL_FLUSH 0
  LOCAL_ADD 1
  LOCAL_CHANGE 2

The function (local_kind) that converts them into strings again
is private. New addons (for example openwrt ubus bindings) that make
use of babeld should also use the same function to convert the kind
to string format. This allows to ensure uniformity over all addons.

Signed-off-by: Nick Hainke <vincent@systemli.org>