Put add_ifcon to configuration.h
If you want to add an interface with a different config you need to call
add_interface with an interface_conf as an argument. add_interface is
already located in the configuration.h. However, add_ifconf is not
available. The commit puts add_ifcon into the header. Further, it also
adds interface_confs to the header. The add_ifconf requires as a
parameter the interface_confs list.

Signed-off-by: Nick Hainke <vincent@systemli.org>