Publish add filter function
Make add_filter public. If you want to add a filter you need to call
add_filter with the corresponding filter list as an argument.
This can be:
- input_filters
- output_filters
- redistribute_filters
- install_filters

This commit puts the add_filter and the filter lists into the header so
filters can be added outside the configuration.c file.

Signed-off-by: Nick Hainke <vincent@systemli.org>