Allow using sha2 and blake2 system libaries
Using ?= to assign LDLIBS and the new BLAKE_* SHA2_* vars allows overriding
them from the environment and use system libraries instead of the bundled
ones.

To make sure includes are picked up from the system instead of the vendor
copies we replace #include "" with #include <> and use -I in CFLAGS.

Finally since keeping SRCS and OBJS in sync is tedious maximus I replace
OBJS with a .c=.o substitution.