Use per-table dumps on kernels where this is available
Starting with version 4.20, the Linux kernel gained the ability to filter
route dumps on (among other things) table ID on the kernel side of the
netlink transaction. This can significantly increase the performance of
route dumps in cases where babeld is configured to only monitor a subset of
the kernel's route tables. In particular, it avoids the per-table lock
contention inside the kernel if another routing daemon is updating a table
that babeld is not using while babeld issues a dump request.

The filtering works by setting the rtm_table attribute on the netlink dump
request, but it only works if the NETLINK_GET_STRICT_CHK socket option is
set on the netlink socket. Older kernels will just ignore the option, so we
can just always pass it to the kernel, and simply break out of the
per-table loop if we detected that table filtering is unavailable.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>