Don't perform full route dumps when routes change.
We used to perform a full route dump whenever the kernel notified
us that the kernel table has changed.  Avoid the dump, and only
install or flush the route that was notified by the kernel.

We still perform periodic route dumps every kernel_check_interval.
This can be set to 0 if periodic route dumps are too costly.